Provides extra space at the bottom of the sheet for mobile device. Fixes #4367
